git-watch: run an arbitrary command on file change
The set of files that are watched are those under revision control with
git. If you are working on a new file, changes on it won't be noticed
unless you mark your intent to commit it with `git add -N` and restart
git-watch.

Requires `entr`, available on Linux and with homebrew.